+++ ttdrest.gemspec
@@ -14,7 +14,7 @@ Gem::Specification.new do |spec|
   spec.license       = "MIT"
 
   #spec.files         = `git ls-files -z`.split("\x0")
-  spec.files         = Dir["README.md","Gemfile","Rakefile", "lib/*", "lib/ttdrest/*", "lib/ttdrest/concerns/*"]
+  spec.files         = Dir["README.md","Gemfile","Rakefile", "lib/*", "lib/ttdrest/*", "lib/ttdrest/concerns/**/*"]
   spec.executables   = spec.files.grep(%r{^bin/}) { |f| File.basename(f) }
   spec.test_files    = spec.files.grep(%r{^(test|spec|features)/})
   spec.require_paths = ["lib"]
